Forgot to commit for a day, oops, lots of shit™

+++ DiscordBot/Rust/Models/Recipe.cs
@@ -10,6 +10,6 @@ namespace DiscordBot.Rust.Models
 
         public ItemCount[] Input;
         public ItemCount Output;
-        public int TTC;
+        public float TTC;
     }
 }
